sql: move ProcessComputedColumns to sqlbase
This brings along a bunch of name resolution stuff too.
