+++ pkg/sql/targets.go
@@ -17,7 +17,6 @@ package sql
 import (
 	"context"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/types"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
@@ -68,7 +67,7 @@ func (p *planner) computeRenderAllowingStars(
 		return nil, nil, false, err
 	}
 
-	if hasStar, cols, typedExprs, err := checkRenderStar(ctx, target, info, ivarHelper); err != nil {
+	if hasStar, cols, typedExprs, err := sqlbase.CheckRenderStar(ctx, target, info, ivarHelper); err != nil {
 		return nil, nil, false, err
 	} else if hasStar {
 		return cols, typedExprs, hasStar, nil
@@ -146,33 +145,4 @@ func (s *renderNode) addOrReuseRenders(
 // string can be used to determine if two expressions are equivalent.
 func symbolicExprStr(expr tree.Expr) string {
 	return tree.AsStringWithFlags(expr, tree.FmtCheckEquivalence)
-}
-
-// checkRenderStar handles the case where the target specification contains a
-// SQL star (UnqualifiedStar or AllColumnsSelector). We match the prefix of the
-// name to one of the tables in the query and then expand the "*" into a list
-// of columns. A sqlbase.ResultColumns and Expr pair is returned for each column.
-func checkRenderStar(
-	ctx context.Context,
-	target tree.SelectExpr,
-	info sqlbase.MultiSourceInfo,
-	ivarHelper tree.IndexedVarHelper,
-) (isStar bool, columns sqlbase.ResultColumns, exprs []tree.TypedExpr, err error) {
-	v, ok := target.Expr.(tree.VarName)
-	if !ok {
-		return false, nil, nil, nil
-	}
-
-	switch v.(type) {
-	case tree.UnqualifiedStar, *tree.AllColumnsSelector:
-		if target.As != "" {
-			return false, nil, nil, pgerror.NewErrorf(pgerror.CodeSyntaxError,
-				"%q cannot be aliased", tree.ErrString(v))
-		}
-
-		columns, exprs, err = expandStar(ctx, info, v, ivarHelper)
-		return true, columns, exprs, err
-	default:
-		return false, nil, nil, nil
-	}
 }
